Refactor how commands are handled
This changes how gram handles commands, specifically settings interactions are scoped under a top level settings

Before this change calling diff would look like 'gram -t token diff-settings ...' with this change 'gram -t token settings diff'.
